client funding functions restrict deposit to client contract owner, this is to allow to manage the allowances more safely
